app/requirements.txt
google-cloud-secret-manager==2.7.1
